azure_host: fix ephemeral OS option
the option used to enable the ephemeral disk as cache for the OS disk.
this commit should fix the behavior and allow to use part of the
ephemeral storage available as main OS disk.
Note that the remaining ephemeral space available will be mapped as a
separate volume and still mounted on /mnt (and so remounted to /data).

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>