k6:Roles&Users: make the resource load more gentle
The module started failing quite deterministically due to either getting
the 204 HTTP Status Code instead of the 200 (which is equally fine) and
also in the user / rolebindings creation part. Seems we have to wait a
bit to create the rolebinding after the user to be more successful.
Moreover... let's keep the virtual users to 1.

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>