terraform/k3d: fix 'node_access_commands' as an empty map
The 'node_access_commands' is returned as a Map/Object from the terraform modules
to map each node to a script to access it.
We return an empty one from k3d, as we don't really provide access
scripts for nodes there, but since the parameter is expected we pass an empty List.
Anyway, to be coherent with the other modules, we should pass an empty Map (otherwise
go unmarshalling complains).

Signed-off-by: Francesco Giudici <francesco.giudici@suse.com>